Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
httpx==0.25.2
